Added flags for partners buttons
