+++ configure.sh
@@ -95,14 +95,19 @@ else
 #define CLOUD_URL ""
 #define MAX_PARTNERS_COUNT 0
 #define PARTNER1_NAME ""
+#define PARTNER1_HAS_BUTTON false
 #define PARTNER1_BANNER_PLACEMENT_ID ""
 #define PARTNER2_NAME ""
+#define PARTNER2_HAS_BUTTON false
 #define PARTNER2_BANNER_PLACEMENT_ID ""
 #define PARTNER3_NAME ""
+#define PARTNER3_HAS_BUTTON false
 #define PARTNER3_BANNER_PLACEMENT_ID ""
 #define PARTNER4_NAME ""
+#define PARTNER4_HAS_BUTTON false
 #define PARTNER4_BANNER_PLACEMENT_ID ""
 #define PARTNER5_NAME ""
+#define PARTNER5_HAS_BUTTON false
 #define PARTNER5_BANNER_PLACEMENT_ID ""
 #define MAXIM_CLIENT_ID ""
 #define MAXIM_SERVER_TOKEN ""
